refactor : GuestsContext 분리
