+++ src/App.tsx
@@ -4,6 +4,7 @@ import theme from "Styles/theme";
 import Reset from "Styles/Reset";
 import Normalize from "Styles/Normalize";
 import Header from "Component/Header/Header";
+import { GuestsProvider } from "Context/GuestsContext";
 
 const App = () => {
 	return (
@@ -11,7 +12,7 @@ const App = () => {
 			<ThemeProvider theme={theme}>
 				<Reset />
 				<Normalize />
-				<Header />
+				<GuestsProvider inner={<Header />} />
 			</ThemeProvider>
 		</div>
 	);
