+++ src/Component/Header/Modal/GuestsDetail.tsx
@@ -1,22 +1,10 @@
-import { createContext, Dispatch, useReducer } from "react";
+import { useContext } from "react";
 
 import { RemoveCircleButton, AddCircleButton } from "util/Icons";
+import { GuestsContext, GuestsDispatchContext } from "Context/GuestsContext";
 import { StyledGuestList, StyledGuestDesc, StyledGuestBtns } from "./Modal.styled";
 
-type guestsType = {
-	adult: number;
-	child: number;
-	baby: number;
-};
-
 type guestType = "adult" | "child" | "baby";
-
-type guestsDispatchAction = {
-	guest: guestType;
-	type: 0 | 1;
-};
-
-type guestsDispatchType = Dispatch<guestsDispatchAction>;
 
 type guestsOptionType = {
 	id: number;
@@ -26,32 +14,10 @@ type guestsOptionType = {
 	state: number;
 };
 
-const guestsDefault = { adult: 0, child: 0, baby: 0 };
-const GuestsContext = createContext<guestsType>(guestsDefault);
-const GuestsDispatchContext = createContext<guestsDispatchType>(() => null);
-
-const guestsReducer = (guests: guestsType, action: guestsDispatchAction) => {
-	const { guest, type } = action;
-	const newGuests = { ...guests };
-
-	switch (type) {
-		case 0:
-			if (newGuests[guest] > 0) newGuests[guest] -= 1;
-			break;
-		case 1:
-			if (newGuests[guest] < 8) newGuests[guest] += 1;
-			break;
-		default:
-	}
-
-	const { adult, child, baby } = newGuests;
-	if (!adult && child + baby) newGuests.adult += 1;
-
-	return newGuests;
-};
-
 const GuestsDatail = () => {
-	const [guests, guestsDispatch] = useReducer(guestsReducer, guestsDefault);
+	const guests = useContext(GuestsContext);
+	const guestsDispatch = useContext(GuestsDispatchContext);
+
 	const { adult, child, baby } = guests;
 
 	const guestsOptions: guestsOptionType[] = [
@@ -87,13 +53,7 @@ const GuestsDatail = () => {
 		);
 	});
 
-	return (
-		<GuestsContext.Provider value={guests}>
-			<GuestsDispatchContext.Provider value={guestsDispatch}>
-				{guestsLists}
-			</GuestsDispatchContext.Provider>
-		</GuestsContext.Provider>
-	);
+	return <div>{guestsLists}</div>;
 };
 
 export default GuestsDatail;
